feat: alert dialog filter expanding
risk: low
