+++ libs/sdk-ui-kit/src/@ui/UiIcon/icons.tsx
@@ -58,4 +58,10 @@ export const iconPaths: Record<IconType, React.ReactNode> = {
     lock: (
         <path d="M15.0049 7.50488C15.415 7.50488 15.7666 7.65137 16.0596 7.94434C16.3525 8.2373 16.499 8.58887 16.499 8.99902V16.9971C16.499 17.4137 16.3525 17.7686 16.0596 18.0615C15.7666 18.3545 15.415 18.501 15.0049 18.501H5.00488C4.58822 18.501 4.2334 18.3545 3.94043 18.0615C3.64746 17.7686 3.50098 17.4137 3.50098 16.9971V8.99902C3.50098 8.58887 3.64746 8.2373 3.94043 7.94434C4.2334 7.65137 4.58822 7.50488 5.00488 7.50488H5.50293V6.00098C5.50293 5.38249 5.62012 4.7998 5.85449 4.25293C6.09538 3.70605 6.41764 3.23079 6.82129 2.82715C7.23145 2.41699 7.70996 2.09473 8.25684 1.86035C8.80371 1.61947 9.38639 1.49902 10.0049 1.49902C10.6234 1.49902 11.2061 1.61947 11.7529 1.86035C12.2998 2.09473 12.7751 2.41699 13.1787 2.82715C13.5889 3.23079 13.9111 3.70605 14.1455 4.25293C14.3864 4.7998 14.5068 5.38249 14.5068 6.00098V7.50488H15.0049ZM6.49902 6.00098V7.50488H13.501V6.00098C13.501 5.51921 13.4098 5.06673 13.2275 4.64355C13.0452 4.21387 12.7946 3.84277 12.4756 3.53027C12.1566 3.21126 11.7855 2.96061 11.3623 2.77832C10.9391 2.59603 10.4867 2.50488 10.0049 2.50488C9.52311 2.50488 9.06738 2.59603 8.6377 2.77832C8.21452 2.96061 7.84342 3.21126 7.52441 3.53027C7.21191 3.84277 6.96126 4.21387 6.77246 4.64355C6.59017 5.06673 6.49902 5.51921 6.49902 6.00098ZM15.5029 16.9971V8.99902C15.5029 8.8623 15.4541 8.74512 15.3564 8.64746C15.2588 8.5498 15.1416 8.50098 15.0049 8.50098H5.00488C4.86816 8.50098 4.75098 8.5498 4.65332 8.64746C4.55566 8.74512 4.50684 8.8623 4.50684 8.99902V16.9971C4.50684 17.1403 4.55566 17.2607 4.65332 17.3584C4.75098 17.4561 4.86816 17.5049 5.00488 17.5049H15.0049C15.1416 17.5049 15.2588 17.4561 15.3564 17.3584C15.4541 17.2607 15.5029 17.1403 15.5029 16.9971Z" />
     ),
+    navigateUp: (
+        <path d="M4.58008 13.5352C4.39779 13.3659 4.30664 13.1576 4.30664 12.9102C4.30664 12.6497 4.4043 12.4284 4.59961 12.2461L9.36523 7.48047C9.45638 7.38932 9.55404 7.32422 9.6582 7.28516C9.77539 7.23307 9.89258 7.20703 10.0098 7.20703C10.14 7.20703 10.2572 7.23307 10.3613 7.28516C10.4785 7.32422 10.5762 7.38932 10.6543 7.48047L15.4395 12.2461C15.5176 12.3372 15.5762 12.4414 15.6152 12.5586C15.6673 12.6628 15.6934 12.7734 15.6934 12.8906C15.6934 13.0078 15.6673 13.125 15.6152 13.2422C15.5632 13.3594 15.498 13.457 15.4199 13.5352C15.3418 13.6263 15.2441 13.6914 15.127 13.7305C15.0098 13.7695 14.8926 13.7891 14.7754 13.7891C14.6582 13.7891 14.541 13.7695 14.4238 13.7305C14.3197 13.6914 14.222 13.6263 14.1309 13.5352L9.99023 9.39453L5.86914 13.5352C5.77799 13.6263 5.67383 13.6914 5.55664 13.7305C5.45247 13.7695 5.3418 13.7891 5.22461 13.7891C5.10742 13.7891 4.99023 13.7695 4.87305 13.7305C4.75586 13.6914 4.6582 13.6263 4.58008 13.5352Z" />
+    ),
+    navigateDown: (
+        <path d="M15.4199 7.48047C15.6022 7.64974 15.6934 7.85807 15.6934 8.10547C15.6934 8.35286 15.5957 8.57422 15.4004 8.76953L10.6152 13.5352C10.5371 13.6263 10.4395 13.6914 10.3223 13.7305C10.2181 13.7695 10.1074 13.7891 9.99023 13.7891C9.86003 13.7891 9.73633 13.7695 9.61914 13.7305C9.51497 13.6914 9.42383 13.6263 9.3457 13.5352L4.56055 8.76953C4.48242 8.67839 4.41732 8.58073 4.36523 8.47656C4.32617 8.35938 4.30664 8.24219 4.30664 8.125C4.30664 7.99479 4.33268 7.8776 4.38477 7.77344C4.43685 7.65625 4.50195 7.55859 4.58008 7.48047C4.6582 7.38932 4.75586 7.32422 4.87305 7.28516C4.99023 7.23307 5.10742 7.20703 5.22461 7.20703C5.3418 7.20703 5.45247 7.23307 5.55664 7.28516C5.67383 7.32422 5.77799 7.38932 5.86914 7.48047L9.99023 11.6016L14.1309 7.48047C14.222 7.38932 14.3197 7.32422 14.4238 7.28516C14.541 7.23307 14.6582 7.20703 14.7754 7.20703C14.8926 7.20703 15.0098 7.23307 15.127 7.28516C15.2441 7.32422 15.3418 7.38932 15.4199 7.48047Z" />
+    ),
 };
